falta limitar las rutas de los usuarios

+++ app/Http/Controllers/PlayersController.php
@@ -6,6 +6,7 @@ use App\Models\Nationality;
 use App\Models\People;
 use Illuminate\Http\Request;
 use App\Models\Player;
+use Illuminate\Support\Facades\Auth;
 
 class PlayersController extends Controller
 {
@@ -14,9 +15,13 @@ class PlayersController extends Controller
      */
     public function index()
     {
-        $nationalities = Nationality::all();
-        $people = People::with('nationality')->get();
-        return view('players.players_view', compact('people', "nationalities"));
+        if (Auth::user() && Auth::user()->rol_id === 1) {
+            $nationalities = Nationality::all();
+            $people = People::with('nationality')->get();
+            return view('players.players_view', compact('people', "nationalities"));
+        }
+
+        return redirect()->route('index');
     }
 
     /**
@@ -24,11 +29,14 @@ class PlayersController extends Controller
      */
     public function create()
     {
-        $nationalities = Nationality::all();
-        $people = People::all();
-        $people = People::with('player')->get();
-        return view('players.create',compact("nationalities","people"));
+        if (Auth::user() && Auth::user()->rol_id === 1) {
+            $nationalities = Nationality::all();
+            $people = People::all();
+            $people = People::with('player')->get();
+            return view('players.create', compact("nationalities", "people"));
+        }
 
+        return redirect()->route('index');
     }
 
     /**
@@ -37,34 +45,35 @@ class PlayersController extends Controller
     public function store(Request $request)
     {
 
+        if (Auth::user() && Auth::user()->rol_id === 1) {
+            $request->validate([
+                "avatar" => "required|image|max:2000"
+            ]);
 
-        $request->validate([
-            "avatar" => "required|image|max:2000"
-        ]);
+            if ($request->hasFile("avatar")) {
+                $avatarPath = $request->file("avatar")->store("avatars", "public");
+            }
 
-        if($request->hasFile("avatar")){
-            $avatarPath = $request->file("avatar")->store("avatars", "public");
+            $people = new People();
+            $people->name = $request->name;
+            $people->lastname = $request->lastname;
+            $people->birthdate = $request->birthdate;
+            $people->birthplace = $request->country;
+            $people->avatar = $avatarPath;
+            $people->save();
+
+
+            $player = new Player();
+            $player->status = $request->status;
+            $player->height = $request->height;
+            $player->bestSide = $request->bestSide;
+            $player->people_id = $people->id;
+            $player->save();
+
+            return redirect()->route("players.index");
         }
 
-        $people = new People();
-        $people->name = $request->name;
-        $people->lastname = $request->lastname;
-        $people->birthdate = $request->birthdate;
-        $people->birthplace = $request->country;
-        $people->avatar = $avatarPath;
-        $people->save();
-
-
-        $player = new Player();
-        $player->status = $request->status;
-        $player->height = $request->height;
-        $player->bestSide=$request->bestSide;
-        $player->people_id = $people->id;
-        $player->save();
-
-        return redirect()->route("players.index");
-
-
+        return redirect()->route('index');
     }
 
     /**
@@ -81,10 +90,13 @@ class PlayersController extends Controller
      */
     public function edit($id)
     {
-        $nationalities = Nationality::all();
-        $player = Player::with('people')->findOrFail($id);
-        $people = People::all();
-        return view('players.edit', compact('player', 'nationalities',"people"));
+        if (Auth::user() && Auth::user()->rol_id === 1) {
+            return redirect()->route('index');
+            $nationalities = Nationality::all();
+            $player = Player::with('people')->findOrFail($id);
+            $people = People::all();
+            return view('players.edit', compact('player', 'nationalities', "people"));
+        }
     }
 
     /**
@@ -92,35 +104,36 @@ class PlayersController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $people = People::findOrFail($id);
-        $player = Player::findOrFail($id);
+        if (Auth::user() && Auth::user()->rol_id === 1) {
+            return redirect()->route('index');
+            $people = People::findOrFail($id);
+            $player = Player::findOrFail($id);
 
-        $request->validate([
-            "avatar" => "nullable|image|max:2000"
-        ]);
-        if($request->hasFile("avatar")){
-            $avatarPath = $request->file("avatar")->store("avatar", "public");
-            $people->avatar = $avatarPath;
+            $request->validate([
+                "avatar" => "nullable|image|max:2000"
+            ]);
+            if ($request->hasFile("avatar")) {
+                $avatarPath = $request->file("avatar")->store("avatar", "public");
+                $people->avatar = $avatarPath;
+            }
+            // CORREGIR SUBIDA DE IMAGEN PLAYERS
 
+            $people = new People();
+            $people->name = $request->name;
+            $people->lastname = $request->lastname;
+            $people->birthdate = $request->birthdate;
+            $people->birthplace = $request->country;
+            $people->save();
+
+
+            $player->status = $request->status;
+            $player->height = $request->height;
+            $player->bestSide = $request->bestSide;
+            $player->people_id = $people->id;
+            $player->save();
+
+            return redirect()->route("players.index");
         }
-        // CORREGIR SUBIDA DE IMAGEN PLAYERS
-
-        $people = new People();
-        $people->name = $request->name;
-        $people->lastname = $request->lastname;
-        $people->birthdate = $request->birthdate;
-        $people->birthplace = $request->country;
-        $people->save();
-
-
-        $player->status = $request->status;
-        $player->height = $request->height;
-        $player->bestSide=$request->bestSide;
-        $player->people_id = $people->id;
-        $player->save();
-
-        return redirect()->route("players.index");
-
     }
 
     /**
@@ -128,12 +141,14 @@ class PlayersController extends Controller
      */
     public function destroy(string $id)
     {
-        $nationalities = Nationality::all();
-        $player = Player::with('people')->findOrFail($id);
-        $player->delete();
-        $person = $player->people;
-        $person->delete();
-        return redirect()->route("players.index");
-
+        if (Auth::user() && Auth::user()->rol_id === 1) {
+            return redirect()->route('index');
+            $nationalities = Nationality::all();
+            $player = Player::with('people')->findOrFail($id);
+            $player->delete();
+            $person = $player->people;
+            $person->delete();
+            return redirect()->route("players.index");
+        }
     }
 }
